+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Unreleased
+## Release 0.2.0 
 
 ### Added